Manage form and error state
